+++ Tests/Functional/Domain/Repository/SpeakerRepositoryTest.php
@@ -1,0 +1,52 @@
+<?php
+
+declare(strict_types=1);
+
+namespace OliverKlee\Seminars\Tests\Functional\Domain\Repository;
+
+use Nimut\TestingFramework\TestCase\FunctionalTestCase;
+use OliverKlee\Seminars\Domain\Model\Speaker;
+use OliverKlee\Seminars\Domain\Repository\SpeakerRepository;
+use TYPO3\CMS\Core\Information\Typo3Version;
+use TYPO3\CMS\Core\Utility\GeneralUtility;
+use TYPO3\CMS\Extbase\Object\ObjectManager;
+
+/**
+ * @covers \OliverKlee\Seminars\Domain\Model\Speaker
+ * @covers \OliverKlee\Seminars\Domain\Repository\SpeakerRepository
+ */
+final class SpeakerRepositoryTest extends FunctionalTestCase
+{
+    protected $testExtensionsToLoad = ['typo3conf/ext/oelib', 'typo3conf/ext/seminars'];
+
+    /**
+     * @var SpeakerRepository
+     */
+    private $subject;
+
+    protected function setUp(): void
+    {
+        parent::setUp();
+
+        if ((new Typo3Version())->getMajorVersion() >= 11) {
+            $this->subject = GeneralUtility::makeInstance(SpeakerRepository::class);
+        } else {
+            $objectManager = GeneralUtility::makeInstance(ObjectManager::class);
+            $this->subject = $objectManager->get(SpeakerRepository::class);
+        }
+    }
+
+    /**
+     * @test
+     */
+    public function mapsAllModelFields(): void
+    {
+        $this->importDataSet(__DIR__ . '/Fixtures/SpeakerRepository/SpeakerWithAllFields.xml');
+
+        $result = $this->subject->findByUid(1);
+
+        self::assertInstanceOf(Speaker::class, $result);
+        self::assertSame('Dan Chase', $result->getName());
+        self::assertSame('dan@example.com', $result->getEmailAddress());
+    }
+}